fix test

+++ subgraph/tests/utils.ts
@@ -13,8 +13,8 @@ export function createNewMetaV1Event(sender: string, subject: Bytes, meta: Bytes
   // Set up transaction data
   metaV1Event.transaction.hash = Bytes.fromHexString(transactionHash);
   metaV1Event.transaction.from = Address.fromString(sender);
-  metaV1Event.block.number = BigInt.fromI32(transactionBlockNumber);
-  metaV1Event.block.timestamp = BigInt.fromI32(transactionTimestamp);
+  metaV1Event.block.number = BigInt.fromI32(transactionBlockNumber as i32);
+  metaV1Event.block.timestamp = BigInt.fromI32(transactionTimestamp as i32);
 
   let senderParam = new ethereum.EventParam("sender", ethereum.Value.fromAddress(Address.fromString(sender)));
   let subjectParam = new ethereum.EventParam("subject", ethereum.Value.fromBytes(subject));
